[PM-40679] Rename desktop_native Autotype Rust code to use MVP in name, separated to its own module.
