+++ apps/desktop/desktop_native/autotype/src/mvp/windows/type_input.rs
@@ -6,7 +6,7 @@ use windows::Win32::UI::Input::KeyboardAndMouse::{
 };
 
 use super::{ErrorOperations, KeyboardShortcutInput, Win32ErrorOperations};
-use crate::get_numeric_modifier_key;
+use crate::mvp::get_numeric_modifier_key;
 
 const IS_VIRTUAL_KEY: bool = true;
 const IS_REAL_KEY: bool = false;
@@ -227,7 +227,7 @@ mod tests {
     use windows::Win32::Foundation::WIN32_ERROR;
 
     use super::*;
-    use crate::windowing::MockErrorOperations;
+    use crate::mvp::windowing::MockErrorOperations;
 
     #[test]
     fn get_alphabetic_hot_key_succeeds() {
