+++ apps/desktop/desktop_native/napi/src/autotype.rs
@@ -2,7 +2,7 @@
 pub mod autotype {
     #[napi]
     pub fn get_foreground_window_title() -> napi::Result<String> {
-        Ok(autotype::get_foreground_window_title()?)
+        Ok(autotype::mvp::get_foreground_window_title()?)
     }
 
     #[napi]
@@ -10,6 +10,6 @@ pub mod autotype {
         input: Vec<u16>,
         keyboard_shortcut: Vec<String>,
     ) -> napi::Result<(), napi::Status> {
-        Ok(autotype::type_input(&input, &keyboard_shortcut)?)
+        Ok(autotype::mvp::type_input(&input, &keyboard_shortcut)?)
     }
 }
